add support for BENCH_FREQ_MHZ to plot.py
When that's specified, the plot becomes a log-log plot of
cycles-per-byte. I've found that automated ways of counting clock cycles
in the benchmarks themselves tend to lie, so it seems like it'll work
better to input the known CPU frequency manually via this env var, and
have the plotting logic do the math. I haven't yet updated plotz.py, but
we can do something similar there.