make derive_key take a key of any length
The previous version of this API called for a key of exactly 256 bits.
That's good for optimal performance, but it would mean losing the
use-with-other-algorithms property for applications whose input keys are
a different size. There's no way for an abstraction over the previous
version to provide reliable domain separation for the "extract" step.

Note also that implementations that want to recover optimal performance
have the option of hardcoding the output words of the first step. I
don't expect this will be at all common in practice, but it's there.